Locale関連の不具合修正
